requirements.txt
transformers
PyPDF2 
sentence-transformers 
faiss-cpu 
streamlit
nltk 